# specs/003-import-from-storygraph-goodreads/contracts/import-api.openapi.yaml
openapi: 3.0.3
info:
  title: Tome Import API
  description: |
    API for importing reading history from TheStoryGraph and Goodreads CSV exports.
    
    **Features:**
    - Upload and parse CSV files from supported providers
    - Intelligent book matching using ISBN, fuzzy title/author matching
    - Preview matched/unmatched records before committing
    - Create reading sessions and sync ratings to Calibre
    - Track unmatched records for manual review
    
    **Workflow:**
    1. POST /api/import/upload - Upload CSV, get preview
    2. GET /api/import/:importId/preview - Review detailed matches
    3. POST /api/import/:importId/execute - Confirm and execute import
    4. GET /api/import/:importId/unmatched - Review unmatched records
  version: 1.0.0
  contact:
    name: Tome Support
    url: https://github.com/yourusername/tome
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://tome.example.com
    description: Production server
...
paths:
...
  /api/import/{importId}/execute:
    post:
      tags:
        - import
      summary: Execute confirmed import
      description: |
        Execute the import after user reviews and confirms matches. This endpoint will:
        - Create reading sessions for matched books
        - Sync ratings to Calibre database
        - Store unmatched records for later review
        - Update import log with final statistics
        
        **Operations Performed:**
        1. Create reading_sessions records (start/end dates from CSV)
        2. Create progress_logs at 100% completion
        3. Update books.rating if different from CSV rating
        4. Call updateCalibreRating() for each rating change
        5. Store unmatched records in import_unmatched_records
        
        **Transaction Behavior:**
        - All session/rating operations in single transaction
        - Rollback on error (except Calibre sync failures)
        - Calibre sync failures logged but don't block import
        
        **Performance:** Processes 1000 records in <30 seconds
      operationId: executeImport
      parameters:
        - name: importId
          in: path
          required: true
          description: UUID of the import to execute
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
            examples:
              confirmAllMatches:
                summary: Confirm all high-confidence matches
                value:
                  confirmedMatches: ["match-1", "match-2", "match-3"]
                  skipRecords: ["unmatched-1", "unmatched-2"]
                  forceDuplicates: false
                  options:
                    syncRatings: true
                    createSessions: true
                    skipExistingSessions: true
              forceAllMatches:
                summary: Force duplicates and confirm all
                value:
                  confirmedMatches: ["match-1", "match-2", "match-3", "match-4"]
                  skipRecords: []
                  forceDuplicates: true
                  options:
                    syncRatings: true
                    createSessions: true
                    skipExistingSessions: false
      responses:
        '200':
          description: Import executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
              examples:
                fullSuccess:
                  summary: All records processed successfully
                  value:
                    success: true
                    summary:
                      sessionsCreated: 220
                      sessionsSkipped: 10
                      ratingsSync: 215
                      calibreSyncFailures: 0
                      unmatchedRecords: 4
                      executionTimeMs: 28450
                    importLogId: 15
                    status: "success"
                    completedAt: "2024-12-01T10:35:28Z"
                partialSuccess:
                  summary: Some Calibre sync failures
                  value:
                    success: true
                    summary:
                      sessionsCreated: 220
                      sessionsSkipped: 10
                      ratingsSync: 210
                      calibreSyncFailures: 5
                      unmatchedRecords: 4
                      executionTimeMs: 29120
                    importLogId: 15
                    status: "partial"
                    completedAt: "2024-12-01T10:35:29Z"
                    warnings:
                      - "Failed to sync 5 ratings to Calibre database"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidMatchIds:
                  summary: Confirmed match IDs don't exist
                  value:
                    success: false
                    error: "Invalid match IDs"
                    details: "Match IDs not found: match-999, match-888"
                    code: "INVALID_MATCH_IDS"
                alreadyExecuted:
                  summary: Import already executed
                  value:
                    success: false
                    error: "Import already executed"
                    details: "This import was already executed on 2024-12-01T11:00:00Z"
                    code: "IMPORT_ALREADY_EXECUTED"
        '404':
          description: Import not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Import execution failed"
                details: "Database transaction failed: UNIQUE constraint failed"
                code: "EXECUTION_FAILED"
...
components:
  schemas:
...
    ExecuteRequest:
      type: object
      required:
        - confirmedMatches
        - skipRecords
        - forceDuplicates
      properties:
        confirmedMatches:
          type: array
          items:
            type: string
          description: Match IDs to process
          example: ["match-1", "match-2", "match-3"]
        skipRecords:
          type: array
          items:
            type: string
          description: Unmatched record IDs to skip
          example: ["unmatched-1", "unmatched-2"]
        forceDuplicates:
          type: boolean
          description: Create sessions even if duplicates exist
          default: false
          example: false
        options:
          type: object
          properties:
            syncRatings:
              type: boolean
              description: Sync ratings to Calibre database
              default: true
            createSessions:
              type: boolean
              description: Create reading sessions
              default: true
            skipExistingSessions:
              type: boolean
              description: Skip duplicate sessions
              default: true
    ExecuteResponse:
      type: object
      required:
        - success
        - summary
        - importLogId
        - status
        - completedAt
      properties:
        success:
          type: boolean
          example: true
        summary:
          $ref: '#/components/schemas/ExecutionSummary'
        importLogId:
          type: integer
          description: Database ID of import_logs record
          example: 15
        status:
          type: string
          enum: [success, partial, failed]
          description: Final import status
          example: "success"
        completedAt:
          type: string
          format: date-time
          example: "2024-12-01T10:35:28Z"
        warnings:
          type: array
          items:
            type: string
          description: Non-fatal warnings
          example: ["Failed to sync 5 ratings to Calibre database"]
    ExecutionSummary:
      type: object
      required:
        - sessionsCreated
        - sessionsSkipped
        - ratingsSync
        - calibreSyncFailures
        - unmatchedRecords
        - executionTimeMs
      properties:
        sessionsCreated:
          type: integer
          minimum: 0
          description: Number of reading sessions created
          example: 220
        sessionsSkipped:
          type: integer
          minimum: 0
          description: Number of duplicate sessions skipped
          example: 10
        ratingsSync:
          type: integer
          minimum: 0
          description: Number of ratings synced to Calibre
          example: 215
        calibreSyncFailures:
          type: integer
          minimum: 0
          description: Number of failed Calibre rating syncs
          example: 0
        unmatchedRecords:
          type: integer
          minimum: 0
          description: Number of unmatched records stored
          example: 4
        executionTimeMs:
          type: integer
          minimum: 0
          description: Total execution time in milliseconds
          example: 28450
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
          example: "Invalid CSV format"
        details:
          type: string
          description: Additional error details
          example: "Missing required columns: Date Read, My Rating"
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_CSV_FORMAT
            - FILE_TOO_LARGE
            - EMPTY_CSV
            - PROVIDER_MISMATCH
            - IMPORT_NOT_FOUND
            - IMPORT_ALREADY_EXECUTED
            - INVALID_MATCH_IDS
            - EXECUTION_FAILED
          example: "INVALID_CSV_FORMAT"
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: Field-specific validation errors
          example:
            file: "File must be a CSV"
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session token (automatically sent by browser)
security:
  - sessionAuth: []